feat(data): extract typed loaders into dsrs-data
DataLoader, typed row mapping, JSONL serialization, and URL detection now live in dsrs-data. The new crate owns the heavy data stack through explicit features so light users do not pay for CSV/Parquet/HuggingFace unless they opt in.

The old dspy-rs data module is reduced to a temporary pass-through, because final test/example relocation still needs a compileable bridge before the aggregator crate is deleted.

Verification:
- cargo check -p dsrs-data --features all
- cargo check --workspace
- cargo test -p dsrs-data --features all
- cargo test --workspace --no-run

Scaffolding: feature gates are in place, but tests still live under dspy-rs until the final hard cutover redistributes them.
