+++ crates/dspy-rs/src/data.rs
@@ -1,0 +1,1 @@
+pub use dsrs_data::*;
