+++ crates/dsrs-data/Cargo.toml
@@ -8,3 +8,26 @@ repository = "https://github.com/krypticmouse/DSRs"
 description = "DSRs dataset loading support."
 
 [dependencies]
+anyhow = "1.0.99"
+bamltype = { path = "../bamltype" }
+csv = { version = "1.3.1", optional = true }
+dsrs-core = { path = "../dsrs-core" }
+dsrs-predict = { path = "../dsrs-predict" }
+parquet = { version = "56.1.0", optional = true }
+arrow = { version = "56.1.0", optional = true }
+hf-hub = { version = "0.4.3", features = ["tokio"], optional = true }
+rayon = "1.10.0"
+regex = "1.11.2"
+reqwest = { version = "0.13", features = ["blocking"], optional = true }
+serde = { version = "1.0.219", features = ["derive"] }
+serde_json = { version = "1.0.140", features = ["preserve_order"] }
+thiserror = "2.0.17"
+tracing = "0.1.44"
+
+[features]
+default = ["csv", "json"]
+all = ["csv", "json", "parquet", "hf-hub"]
+csv = ["dep:csv"]
+json = []
+parquet = ["dep:arrow", "dep:parquet"]
+hf-hub = ["dep:hf-hub", "dep:reqwest", "parquet"]
